requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
Werkzeug==1.0.1
WTForms==2.3.3
numpy
scipy
scikit-learn
pandas
